untitled folder 5/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
httpx==0.26.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0
python-multipart==0.0.6
pygithub==2.2.0
openai==1.6.1
langchain==0.1.10
tiktoken==0.5.2
pgvector==0.2.3
numpy==1.26.3
pyyaml==6.0.1
pytest==7.4.3
tenacity==8.2.3
streamlit==1.32.0
markdown==3.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
transformers==4.34.0
torch==2.0.1
sentencepiece==0.1.99
accelerate==0.23.0
